qr: add another encoding example
This pulls an example from the ISO standard. The test cases uses a
different result to the example in the standard because the
implementation currently calculates the penalty differently.

I've opted not to try and fix this yet as this is intended only to
provide a better baseline for the current implementation.